Add files via upload
The examples that were previously uploaded are getting edited to be automatic. The automated homography finds specific shapes in the corner of the terrain allowing the corners to be determined and transformed. Also automatic shape detection for the hexagons within the terrain allowing us to determine the HSV pixel values.